Adding delete current user account feature

+++ identity_service/identity_service/Controllers/IdentityController.cs
@@ -3,6 +3,7 @@ using Models.Auth;
 using Models.Exceptions;
 using Models.Identities;
 using Models.Services_Interfaces;
+using System.IdentityModel.Tokens.Jwt;
 
 namespace identity_service.Controllers
 {
@@ -56,6 +57,36 @@ namespace identity_service.Controllers
         {
             try
             {
+
+                identityService.DeleteIdentity(email.ToLower());
+                return NoContent();
+            }
+            catch (AssetAlreadyExistException)
+            {
+                return BadRequest("The email is already is been used.");
+            }
+            catch (Exception)
+            {
+                return StatusCode(500);
+            }
+        }
+
+
+        [HttpDelete("me")]
+        public IActionResult Delete()
+        {
+            try
+            {
+                string authorizationHeader = Request.Headers["Authorization"];
+                if (string.IsNullOrEmpty(authorizationHeader) || !authorizationHeader.StartsWith("Bearer "))
+                {
+                    return BadRequest("Invalid or missing Authorization header");
+                }
+                string token = authorizationHeader.Substring("Bearer ".Length).Trim();
+
+                var tokenHandler = new JwtSecurityTokenHandler();
+                var decodedToken = tokenHandler.ReadJwtToken(token);
+                var email = decodedToken.Claims.FirstOrDefault(c => c.Type == "Email")?.Value;
 
                 identityService.DeleteIdentity(email.ToLower());
                 return NoContent();
